[BrowserSession] unifie la gestion du navigateur
